python-example/requirements.txt
paho-mqtt
black